move structure parsing to separate module to avoid jsonschema dep for usage
